backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pytesseract
Pillow
reportlab
